changed order of installing refs (#327)
* changed order of installing refs

* Add EnsureOutputFolder method and integrate it into EnsureCsProjFile

Added a new private method `EnsureOutputFolder` that checks if the
output folder specified by `AxSharpProject.OutputFolder` exists, and
if not, creates it. This method then returns the path to the output
folder.

Modified the `EnsureCsProjFile` method to call the newly added
`EnsureOutputFolder` method. This ensures that the output folder is
created before proceeding with the rest of the method's logic,
preventing potential errors related to missing directories.

---------

Co-authored-by: blazej.kuhajda <blazej.kuhajda@mts.sk>
Co-authored-by: PTKu <61538034+PTKu@users.noreply.github.com>