Add default TargetPlatformMoniker and new CompilerTestOptions
Added a check in AXSharpProject.cs for the TargetPlatfromMoniker
property of CompilerOptions, setting a default value of "ax" and
logging a warning if it is null. Updated IxProjectTests.cs to
include a new using directive for AXSharp.CompilerTests and
modified the AXSharpProject constructor call to include two
instances of CompilerTestOptions. Added a new file,
CompilerTestOptions.cs, defining the CompilerTestOptions class
with default values and basic getter and setter implementations.