Handle InitValue formatting in AddedPropertySetterAstNode
Modified the Init method of AddedPropertySetterAstNode to handle
InitValue differently based on its content. The text from
AddedPropertyInitializer is first assigned to a variable
asFreeTextLiteral. If asFreeTextLiteral starts with a double quote,
InitValue is prefixed with an @ symbol. Otherwise, InitValue is
assigned the value of asFreeTextLiteral without modification. This
ensures correct formatting of InitValue.